Add some copy-paste skeleton of an upscaler idea

+++ src/las_data.rs
@@ -7,11 +7,11 @@ use walkdir::WalkDir;
 
 pub fn load_from_directory<F>(
     path: &str,
-    (scale_x, scale_y, scale_z): (f64, f64, f64),
+    (scale_x, scale_y, scale_z): (f32, f32, f32),
     max_threads: usize,
     mut f: F,
 ) where
-    F: FnMut(f64, f64, f64) -> (),
+    F: FnMut(f32, f32, f32) -> (),
 {
     info!("Beginning iteration over all LAS data");
 
@@ -35,9 +35,9 @@ pub fn load_from_directory<F>(
                 let wrapped_point = wrapped_point.unwrap();
 
                 let (x, y, z) = (
-                    wrapped_point.x * scale_x,
-                    wrapped_point.y * scale_y,
-                    wrapped_point.z * scale_z,
+                    wrapped_point.x as f32 * scale_x,
+                    wrapped_point.y as f32 * scale_y,
+                    wrapped_point.z as f32 * scale_z,
                 );
 
                 sender.send((x, y, z)).unwrap();
@@ -57,22 +57,22 @@ pub fn load_from_directory<F>(
 
 #[derive(Debug, Default, Clone)]
 pub struct Limits {
-    pub min_x: f64,
-    pub min_y: f64,
-    pub min_z: f64,
-    pub max_x: f64,
-    pub max_y: f64,
-    pub max_z: f64,
+    pub min_x: f32,
+    pub min_y: f32,
+    pub min_z: f32,
+    pub max_x: f32,
+    pub max_y: f32,
+    pub max_z: f32,
 }
 
 impl Limits {
-    pub fn load_from_directory(path: &str, scalers: (f64, f64, f64), max_threads: usize) -> Self {
-        let mut max_x: Option<f64> = None;
-        let mut min_x: Option<f64> = None;
-        let mut max_z: Option<f64> = None;
-        let mut min_z: Option<f64> = None;
-        let mut max_y: Option<f64> = None;
-        let mut min_y: Option<f64> = None;
+    pub fn load_from_directory(path: &str, scalers: (f32, f32, f32), max_threads: usize) -> Self {
+        let mut max_x: Option<f32> = None;
+        let mut min_x: Option<f32> = None;
+        let mut max_z: Option<f32> = None;
+        let mut min_z: Option<f32> = None;
+        let mut max_y: Option<f32> = None;
+        let mut min_y: Option<f32> = None;
         load_from_directory(path, scalers, max_threads, |x, y, z| {
             max_x = Some(max_x.unwrap_or(x).max(x));
             max_y = Some(max_y.unwrap_or(y).max(y));
